+++ library-base/src/main/java/com/xtree/base/vo/TopSpeedDomain.java
@@ -5,7 +5,7 @@ import android.os.Parcel;
 public class TopSpeedDomain implements BaseBean, Comparable<TopSpeedDomain> {
     public String url;
     public long speedSec;
-    public float curCTSSec = 0;
+    public long curCTSSec = 0;
 
     @Override
     public int describeContents() {
@@ -16,13 +16,13 @@ public class TopSpeedDomain implements BaseBean, Comparable<TopSpeedDomain> {
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeString(this.url);
         dest.writeLong(this.speedSec);
-        dest.writeFloat(this.curCTSSec);
+        dest.writeLong(this.curCTSSec);
     }
 
     public void readFromParcel(Parcel source) {
         this.url = source.readString();
         this.speedSec = source.readLong();
-        this.curCTSSec = source.readFloat();
+        this.curCTSSec = source.readLong();
     }
 
     public TopSpeedDomain() {
@@ -31,7 +31,7 @@ public class TopSpeedDomain implements BaseBean, Comparable<TopSpeedDomain> {
     protected TopSpeedDomain(Parcel in) {
         this.url = in.readString();
         this.speedSec = in.readLong();
-        this.curCTSSec = in.readFloat();
+        this.curCTSSec = in.readLong();
     }
 
     public static final Creator<TopSpeedDomain> CREATOR = new Creator<TopSpeedDomain>() {
